Removed all legacy graphing code and dependancies
