+++ app/build.gradle
@@ -48,19 +48,19 @@ dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     //implementation 'com.google.android.gms:play-services-maps:+'
     //implementation 'com.google.android.gms:play-services-location:+'
-    implementation 'com.google.android.gms:play-services-maps:16.0.0'
-    implementation 'com.google.android.gms:play-services-location:16.0.0'
+    implementation 'com.google.android.gms:play-services-maps:17.0.0'
+    implementation 'com.google.android.gms:play-services-location:17.0.0'
 
-    implementation 'com.github.PhilJay:MPAndroidChart:v3.0.0'
-    implementation 'com.getbase:floatingactionbutton:+'
+//    implementation 'com.github.PhilJay:MPAndroidChart:v3.0.0'
+    implementation 'com.getbase:floatingactionbutton:1.10.1'
 
-    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
 //    implementation 'com.android.support:appcompat:29.0.2'
     implementation 'androidx.legacy:legacy-preference-v14:1.0.0'
-    implementation 'com.google.android.material:material:1.0.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.material:material:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.1'
 
-    implementation 'org.apache.commons:commons-math3:+'
+    implementation 'org.apache.commons:commons-math3:3.6.1'
 
-    implementation 'org.greenrobot:eventbus:+'
+    implementation 'org.greenrobot:eventbus:3.2.0'
 }
