+++ app/src/main/res/values/strings.xml
@@ -10,6 +10,4 @@
     <string name="TimerPlaceholder">##:##</string>
     <string name="PlaceholderFileSize">##kb</string>
     <string name="loggingLabel">Logging:</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
 </resources>
