+++ app/src/release/res/values/google_maps_api.xml
@@ -1,18 +1,16 @@
-<resources>
-	<!--
-    TODO: Before you run your application, you need a Google Maps API key.
+<!--<resources>-->
+<!--	&lt;!&ndash;-->
+<!--    To get one, follow this link, follow the directions and press "Create" at the end:-->
 
-    To get one, follow this link, follow the directions and press "Create" at the end:
+<!--https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=67:0D:57:71:FA:05:C8:EE:EB:07:A4:C2:72:0C:63:CC:7D:6A:78:6D%3Bcom.driven.rowan.drivenbluetooth-->
 
-https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=67:0D:57:71:FA:05:C8:EE:EB:07:A4:C2:72:0C:63:CC:7D:6A:78:6D%3Bcom.driven.rowan.drivenbluetooth
+<!--    You can also add your credentials to an existing key, using this line:-->
+<!--    67:0D:57:71:FA:05:C8:EE:EB:07:A4:C2:72:0C:63:CC:7D:6A:78:6D;com.driven.rowan.drivenbluetooth-->
 
-    You can also add your credentials to an existing key, using this line:
-    67:0D:57:71:FA:05:C8:EE:EB:07:A4:C2:72:0C:63:CC:7D:6A:78:6D;com.driven.rowan.drivenbluetooth
-
-    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
-    string in this file.
-    -->
-	<string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">
-		YOUR_KEY_HERE
-	</string>
-</resources>
+<!--    Once you have your key (it starts with "AIza"), replace the "google_maps_key"-->
+<!--    string in this file.-->
+<!--    &ndash;&gt;-->
+<!--	<string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">-->
+<!--		YOUR_KEY_HERE-->
+<!--	</string>-->
+<!--</resources>-->
